requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
Faker==40.13.0
pytest==9.0.3
pytest-cov==7.1.0
requests==2.33.1
selenium==4.41.0
seletools==1.5.0
